# openapi.yaml
openapi: 3.0.3
info:
  title: AdrenRush API
  version: "1.0.0"
  description: >
    REST API сайта рейтинга энергетиков **AdrenRush**.

    Аутентификация — JWT в заголовке `Authorization: Bearer <token>`.
    Токен выдаётся при регистрации/входе (`/api/auth/register`, `/api/auth/login`).

    Уровни доступа (см. `SecurityConfig`):
    - `GET /api/drinks/**`, `GET /api/users/**`, `GET /media/**`, `/api/auth/**` — без авторизации;
    - `/api/admin/**` — только роль `ADMIN`;
    - изменяющие операции над напитками/фото — требуют токен и проверяются на роль `ADMIN` в контроллере (иначе `403`);
    - остальные эндпоинты требуют валидный токен.

    Ошибки бизнес-логики возвращаются как `{"error": "..."}` с соответствующим HTTP-статусом.
servers:
  - url: http://localhost:8080
    description: Локальная разработка
  - url: https://adrenrush.ru
    description: Боевой сервер
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Регистрация нового пользователя
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Пользователь создан, выдан токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Вход по логину и паролю
      description: >
        При включённой 2FA, если `code` не передан или неверен, возвращается `401`
        с телом `{ "error": "...", "requires2fa": true }`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход, выдан токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Неверные учётные данные либо требуется 2FA
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/TwoFactorError'
  /api/auth/recover:
    post:
      tags: [Auth]
      summary: Восстановление пароля по коду из приложения-аутентификатора
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecoverRequest'
      responses:
        '200':
          description: Пароль изменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Текущий пользователь
      responses:
        '200':
          description: Данные текущего пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Auth]
      summary: Обновить отображаемое имя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me/avatar:
    post:
      tags: [Auth]
      summary: Загрузить аватар
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Файл изображения (image/*)
      responses:
        '200':
          description: Профиль с новым аватаром
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '507':
          description: Не удалось сохранить файл в хранилище
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me/password:
    post:
      tags: [Auth]
      summary: Сменить пароль
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Пароль изменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me/privacy:
    post:
      tags: [Auth]
      summary: Включить/выключить приватность профиля
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyRequest'
      responses:
        '200':
          description: Обновлённый профиль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me/email:
    get:
      tags: [Auth]
      summary: Состояние привязки почты
      description: |
        Привязанная почта и незавершённая заявка на подтверждение.
        Почта отдаётся только владельцу — в публичном профиле (`GET /api/users/{id}`) её нет.
      responses:
        '200':
          description: Состояние привязки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me/email/request:
    post:
      tags: [Auth]
      summary: Выслать код подтверждения на почту
      description: |
        Шаг 1 привязки. Код из 6 цифр действует 15 минут, следующее письмо —
        не раньше чем через 60 секунд (пауза общая, смена адреса её не сбрасывает).
        Одна почта может быть привязана только к одному аккаунту.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '200':
          description: Письмо отправлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailPending'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Почта занята другим аккаунтом или подтверждается им прямо сейчас
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me/email/confirm:
    post:
      tags: [Auth]
      summary: Подтвердить код и привязать почту
      description: Шаг 2 привязки. Не более 5 попыток ввода, дальше нужен новый код.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailConfirmRequest'
      responses:
        '200':
          description: Почта привязана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Почту успели привязать к другому аккаунту
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me/email/cancel:
    post:
      tags: [Auth]
      summary: Отменить незавершённую привязку
      responses:
        '200':
          description: Заявка отменена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me/email/unbind:
    post:
      tags: [Auth]
      summary: Отвязать почту от аккаунта
      description: Требует пароль от аккаунта. После отвязки адрес снова свободен.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailUnbindRequest'
      responses:
        '200':
          description: Почта отвязана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/2fa/setup:
    post:
      tags: [Auth]
      summary: Сгенерировать секрет 2FA (TOTP)
      description: Возвращает секрет, otpauth-ссылку и QR в виде data-URL. Подтверждается отдельным шагом `/2fa/enable`.
      responses:
        '200':
          description: Данные для настройки приложения-аутентификатора
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TwoFactorSetupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/2fa/enable:
    post:
      tags: [Auth]
      summary: Подтвердить и включить 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorCodeRequest'
      responses:
        '200':
          description: 2FA включена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example: { status: enabled }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/2fa/disable:
    post:
      tags: [Auth]
      summary: Отключить 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoFactorCodeRequest'
      responses:
        '200':
          description: 2FA отключена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example: { status: disabled }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ────────────────────────────── USERS ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Требуется авторизация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Конфликт (например, имя пользователя занято)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
      example: { error: "Сообщение об ошибке" }
    TwoFactorError:
      type: object
      description: Возвращается при входе, когда требуется код 2FA
      properties:
        error:
          type: string
        requires2fa:
          type: boolean
          example: true
      required: [error, requires2fa]
...
    TokenResponse:
      type: object
      properties:
        token:
          type: string
      required: [token]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
      required: [status]
    TwoFactorSetupResponse:
      type: object
      properties:
        secret:
          type: string
        otpauthUrl:
          type: string
        qrDataUrl:
          type: string
          description: QR-код в виде data-URL (base64 PNG)
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/RoleEnum'
        avatarUrl:
          type: string
          nullable: true
        profilePrivate:
          type: boolean
        totpEnabled:
          type: boolean
...
    RoleEnum:
      type: string
      enum: [USER, ADMIN, BANNED]
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fingerprint:
          type: string
          nullable: true
          description: Отпечаток устройства (для связывания аккаунтов)
      required: [username, password]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        code:
          type: string
          nullable: true
          description: Код 2FA, если включена
        fingerprint:
          type: string
          nullable: true
      required: [username, password]
    RecoverRequest:
      type: object
      properties:
        username:
          type: string
        code:
          type: string
          description: Код из приложения-аутентификатора
        newPassword:
          type: string
      required: [username, code, newPassword]
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
          minLength: 4
      required: [oldPassword, newPassword]
    PrivacyRequest:
      type: object
      properties:
        private:
          type: boolean
      required: [private]
    TwoFactorCodeRequest:
      type: object
      properties:
        code:
          type: string
      required: [code]
    EmailRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Приводится к нижнему регистру и обрезается по краям.
          example: you@example.com
      required: [email]
    EmailConfirmRequest:
      type: object
      properties:
        code:
          type: string
          description: 6 цифр из письма.
          example: "247512"
      required: [code]
    EmailUnbindRequest:
      type: object
      properties:
        password:
          type: string
          description: Пароль от аккаунта — подтверждает отвязку.
      required: [password]
    EmailPending:
      type: object
      description: Незавершённая заявка на привязку.
      properties:
        email:
          type: string
          format: email
        expiresInSeconds:
          type: integer
          description: Сколько ещё действует код.
          example: 899
        resendInSeconds:
          type: integer
          description: Сколько ждать до следующего письма (0 — можно отправлять).
          example: 59
        attemptsLeft:
          type: integer
          description: Сколько осталось попыток ввода кода.
          example: 5
        delivered:
          type: boolean
          description: |
            Только в ответе `/request`. `false` — SMTP на сервере не настроен,
            письмо не ушло, код записан в лог приложения.
      required: [email, expiresInSeconds, resendInSeconds, attemptsLeft]
    EmailStatus:
      type: object
      properties:
        email:
          type: string
          format: email
          nullable: true
          description: Привязанная почта; `null` — не привязана.
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        pending:
          allOf:
            - $ref: '#/components/schemas/EmailPending'
          nullable: true
          description: Незавершённая заявка либо `null`.
      required: [email, verifiedAt, pending]
...